feat(pricing): add metadata and refactor component structure
